LPS-158259 Refactor to simplify the GraphQLServletExtender class by implementing DataFetchers for GraphQL fields and creating Processors
